* commit '743855202d0a59078cfb2119911906d0bb094321':
  bug #PSBM-41816

+++ Libraries/Transponster/domain_data.h
@@ -880,7 +880,7 @@ namespace Xml
 {
 struct PPortNumber
 {
-	typedef qint16 value_type;
+	typedef qint32 value_type;
 };
 
 } // namespace Xml
